+++ src/main/java/com/techeer/cokkiri/domain/study/repository/StudyRepository.java
@@ -1,0 +1,6 @@
+package com.techeer.cokkiri.domain.study.repository;
+
+import com.techeer.cokkiri.domain.study.entity.Study;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface StudyRepository extends JpaRepository<Study, Long> {}
